# microservice/source/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Inventory API
  description: |-
    A demo inventory management system.
  version: 2.0.0
servers:
  - url: https://inventory-api.igraphql.co/api/v2
    description: Microservice
...
paths:
...
  /price/{skuId}:
    post:
      tags:
        - Price
      description: Creates a product price.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePrice'
      responses:
        "201":
          description: Price successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode201"
        "400":
          description: Client sent an invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode400"
        "404":
          description: SKU not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "409":
          description: Price already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode409"
        "415":
          description: Invalid Content-Type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode415"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    get:
      tags:
        - Price
      description: Retrieves a product price.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
        - $ref: '#/components/parameters/Field'
      responses:
        "200":
          description: Price successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPrice"
        "404":
          description: Price or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    put:
      tags:
        - Price
      description: Updates a product price.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePrice'
      responses:
        "200":
          description: Price successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "400":
          description: Client sent an invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode400"
        "404":
          description: Price or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "415":
          description: Invalid Content-Type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode415"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
    delete:
      tags:
        - Price
      description: Deletes a product price.
      parameters:
        - name: skuId
          in: path
          description: Product SKU.
          required: true
          schema:
            type: integer
            example: 96070
      responses:
        "200":
          description: Price successfully deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode200"
        "404":
          description: Price or SKU not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode404"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpCode500"
...
components:
  parameters:
...
    Field:
      name: field
      in: query
      description: Limit the fields returned by the query. Can be specified multiple times.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
...
  schemas:
...
    CreatePrice:
      type: object
      required:
        - retail
        - sale
      properties:
        retail:
          type: number
          example: 84.99
        sale:
          type: number
          example: null
    GetPrice:
      type: object
      properties:
        created:
          type: string
          example: 2023-07-04T12:00:00.00.000Z
        retail:
          type: number
          example: 84.99
        sale:
          type: number
          example: null
        sku_id:
          type: integer
          example: 96070
        updated:
          type: string
          example: 2023-07-04T12:00:00.00.000Z
    UpdatePrice:
      type: object
      properties:
        retail:
          type: number
          example: 84.99
        sale:
          type: number
          example: null
...
    HttpCode200:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: ok
    HttpCode201:
      type: object
      properties:
        status:
          type: integer
          example: 201
        message:
          type: string
          example: created
    HttpCode400:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: bad request
    HttpCode404:
      type: object
      properties:
        status:
          type: integer
          example: 404
        message:
          type: string
          example: not found
    HttpCode409:
      type: object
      properties:
        status:
          type: integer
          example: 409
        message:
          type: string
          example: conflict
    HttpCode415:
      type: object
      properties:
        status:
          type: integer
          example: 415
        message:
          type: string
          example: unsupported media type
    HttpCode500:
      type: object
      properties:
        status:
          type: integer
          example: 500
        message:
          type: string
          example: internal server error